requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==4.2.0
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==21.4.0
Authlib==1.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
bitarray==2.5.1
black==23.9.1
bleach==5.0.1
blinker==1.6.3
cachetools==5.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
ci-info==0.2.0
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
configparser==5.2.0
cryptography==41.0.4
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
etelemetry==0.3.0
Events==0.4
exceptiongroup==1.1.3
executing==0.8.3
fastapi==0.103.2
fastjsonschema==2.15.3
filelock==3.7.1
fitz-douyu==0.0.1
fitz-message==0.2
Flask==3.0.0
frontend==0.0.3
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.37
h11==0.13.0
httplib2==0.20.4
huggingface-hub==0.17.3
idna==3.3
importlib-metadata==6.8.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
looseversion==1.0.1
lxml==4.9.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.3.0
multidict==6.0.4
mypy==1.5.1
mypy-extensions==1.0.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.8
networkx==2.8.4
nibabel==4.0.1
nipype==1.8.2
nltk==3.8.1
notebook==6.4.12
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.11.2
pdfreader==0.1.11
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.3
prov==2.0.0
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.7.1
pydot==1.4.2
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==2.5.0
PyPDF4==1.27.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytils==0.4.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.6
pyxnat==1.4
PyYAML==6.0.1
pyzmq==23.2.0
rdflib==6.1.1
regex==2023.10.3
requests==2.31.0
rich==12.5.1
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
starlette==0.27.0
streamlit==1.11.0
sympy==1.12
terminado==0.15.0
threadpoolctl==3.2.0
tinycss2==1.1.1
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
tools==0.1.9
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.2
tqdm==4.66.1
traitlets==5.3.0
traits==6.3.2
transformers==4.34.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2022.1
tzlocal==4.2
urllib3==2.0.6
uvicorn==0.23.2
validators==0.22.0
wasabi==1.1.2
watchdog==2.1.9
wcwidth==0.2.5
weaviate-client==3.24.2
webencodings==0.5.1
Werkzeug==3.0.0
widgetsnbextension==3.6.1
yarl==1.9.2
zipp==3.8.1
